Update README.md
This is a simple GUI based project which is very easy to understand and
use. Talking about the system, it contains all the basic functions which
include adding, viewing, deleting and updating items. In order to add a
medicine detail, the user has to provide its name, price, quantity,
category, and discount.
This simple GUI based Pharmacy Management system provides the
simplest management of medicine items. In short, this project mainly
focuses on CRUD. There’s an external database connection file used in
this mini project to save user’s data permanently